:recycle:[refactor]  공용 컴포넌트 정리
