+++ src/app/components/admin/common/DeleteModal.tsx
@@ -7,7 +7,6 @@ interface DeleteModalProps {
 const DeleteModal = ({ closeModal, deleteUser, text }: DeleteModalProps) => {
   const handleDelete = () => {
     if (deleteUser) {
-      console.log('삭제 완료')
       deleteUser()
       closeModal()
     }
