--- src/app/components/admin/common/mookData.ts
@@ -1,171 +1,0 @@
-export const mookUser = [
-  {
-    memberId: 1,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 2,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 3,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 4,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 5,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 6,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 7,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 8,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-  {
-    memberId: 9,
-    registrationDate: '2024-01-01',
-    email: 'member1@gahcon.ac.kr',
-    grade: '4',
-    major: '컴퓨터공학과',
-    jobKeyword: '웹 개발',
-    point: 1000,
-  },
-]
-
-export const mookBoard = [
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-  {
-    createdAt: '2024-04-04',
-    memberName: 'yng1404',
-    boardId: 1,
-    title: '취업을 하고 싶어',
-    type: '취업 정보',
-  },
-]
-export const mookModelManage = [{}]
